CUT `UploadTestCase`
